DEB: remove polkit rule in the old pkla format
Starting with Ubuntu 25.04 and the future Debian 13, this format is no longer supported. We will follow what
is already done when packaging the RPM and only use the new format, recommending polkitd as a dependency (we
cannot require it, otherwise it will break the installation on older versions). The sleep prevention
functionality will work on Debian 12 and Ubuntu 24.04 and later.

Closes: https://github.com/FoldingAtHome/fah-client-bastet/issues/342