Set service stop timeout to 60 seconds
Although the default is usually higher, some distros and users configure their default service stop timeout to be much lower. If the service is stopped and this much time passes before the client exits, it is killed forcefully. Setting the stop timeout explicitly is good practice and prevents users from losing WUs if the default timeout is low.